models/postにcontrollerの処理を追記（getPostをgetUserPostへ変更・getPost追記）（develop）
